Improvements to ZK error analysis script
Change tau to r_succ since r_succ is the ultimate value that
determines ZK error and soundness error.
Reuse code in PineValid to compute the number of bits needed for
wraparound checks in find_wr_params.sage script.
Print parameters into a table, so it's easier to view and we may
put it into the draft.