Clean up variable names, add comments, and fix a minor bug
Add comments with references to the paper and align variable names more
with the paper.

Also, fix a minor bug in `encode_wr_checks()` where the Client
conditionally sets `wr_check_v` to 0 if the success bit is set. The
Aggregators only have a share of this bit and so cannot set `wr_check_v`
correctly. Instead, set `wr_check_v` to the same value regardless of
whether the check succeeds.