Begin pass of draft
I'm working my way through the draft and cleaning as I go. The goal for
the first pass is to flesh out the skeleton. There are already lots of
bits we can use here.

One thing I'm doing is de-emphasizing the "completeness problem" for
PINE. Given how we will choose parameters, it seems to me that the
impact on ZK is really going to be negligible.

Co-authored-by: Junye Chen <junyec@apple.com>